Brought out start command to separate file
