+++ bot/commands/__init__.py
@@ -1,4 +1,7 @@
 from .request import command_request
 from .avaliable import command_avaliable
 from .qa import command_qa
-__ALL__ = ['command_request', 'command_avaliable','qa']
+from .start import command_start
+
+__ALL__ = ['command_request', 'command_avaliable','qa',
+           'command_start']
